# [goalc] Cleaned up speedups