… common/goos/Object.h
… common/goos/Reader.cpp
… common/serialization/subtitles/subtitles.cpp
… common/serialization/text/text_ser.cpp
… common/type_system/defenum.cpp
… common/type_system/deftype.cpp
… common/type_system/deftype.h
… common/util/crc32.cpp
… common/util/crc32.h
… decompiler/IR2/FormExpressionAnalysis.cpp
… decompiler/analysis/insert_lets.cpp
… decompiler/util/DecompilerTypeSystem.cpp
… decompiler/util/data_decompile.cpp
… goal_src/goos-lib.gs
… goalc/compiler/Compiler.cpp
… goalc/compiler/Compiler.h
… goalc/compiler/CompilerSettings.cpp
… goalc/compiler/Env.cpp
… goalc/compiler/Env.h
… goalc/compiler/Lambda.h
… goalc/compiler/Util.cpp
… goalc/compiler/compilation/Asm.cpp
… goalc/compiler/compilation/Atoms.cpp
… goalc/compiler/compilation/CompilerControl.cpp
… goalc/compiler/compilation/ConstantPropagation.cpp
… goalc/compiler/compilation/ControlFlow.cpp
… goalc/compiler/compilation/Define.cpp
… goalc/compiler/compilation/Function.cpp
… goalc/compiler/compilation/Macro.cpp
… goalc/compiler/compilation/State.cpp